product_service/requirements.txt
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
cloudinary
daphne
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar
email_validator==2.2.0
flower==2.0.1
gunicorn==23.0.0
pymongo==4.11.1
pytest==8.3.3
pytest-asyncio
pytest-django==4.9.0
python-decouple==3.8
sentry-sdk==2.15.0
requests==2.32.3
whitenoise==6.7.0
